LayoutTests/imported/w3c/web-platform-tests/web-animations/interfaces/Animation/cancel-event-on-element-removal-expected.txt


PASS Animation.cancel event fires when element is removed
